+++ app.js
@@ -1,0 +1,4 @@
+const {Probot} = require('probot');
+const {app} = require('@graphql-inspector/github');
+
+Probot.run(app);