+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - **github**: put annotations of removed parts of schema on new schema [#1414](https://github.com/kamilkisiela/graphql-inspector/issues/1414)
 - **ci**: improve UX [#1413](https://github.com/kamilkisiela/graphql-inspector/pull/1413)
+- **github** schema change notifications on Slack, Discord and through Webhook.
+- **github** **action** opt-out annotations
+- **github** **action** opt-in forced success on breaking changes
 
 ### v1.30.4